docs(README): initialize README with "coming soon" info
This is a new repository, and work is still being done on the initial
video element stuff, so it may take some time before this is totally ready.

It's fun to work on though, and judging from some talk on IRC, it seems like
it may be useful to some people.